committing 0x0F-python-object_relational_mapping

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,34 +1,25 @@
 #!/usr/bin/python3
+"""0x0F. Python - Object-relational mapping - task 13. Delete states
 """
-    A script that deletes all State objects from hbtn_0e_6_usa that conatin
-    the letter a
-    Username, password and dbname wil be passed as arguments to the script.
-"""
-
-
-import sys
-from model_state import Base, State
-from sqlalchemy.orm import sessionmaker
-from sqlalchemy import create_engine
 
 if __name__ == '__main__':
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/{}'.format(
-                           sys.argv[1], sys.argv[2], sys.argv[3]),
+    from sys import argv, exit
+    from sqlalchemy import create_engine
+    from sqlalchemy.orm import Session
+    from model_state import Base, State
+
+    if len(argv) != 4:
+        exit('Use: 13-model_state_delete_a.py <mysql username> '
+             '<mysql password> <database name> ')
+
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/'
+                           '{}'.format(argv[1], argv[2], argv[3]),
                            pool_pre_ping=True)
+    session = Session(engine)
+    Base.metadata.create_all(engine)  # creates decprecated warning 1681
 
-    Session = sessionmaker(bind=engine)
-    Base.metadata.create_all(engine)
-
-    # create a session
-    session = Session()
-
-    # extract states with a in them
-    states = session.query(State).filter(State.name.ilike('%a%')).all()
-
-    # delete states
-    for state in states:
+    a_states = session.query(State).filter(State.name.like('%a%')).all()
+    for state in a_states:
         session.delete(state)
-
     session.commit()
-
     session.close()
